+ Added Product Details Screen + Added Token Interceptor

+++ data/src/main/java/com/route/data/model/products/ProductDetailsResponseDM.kt
@@ -4,69 +4,69 @@ import com.google.gson.annotations.SerializedName
 
 data class ProductDetailsResponseDM(
 
-	@field:SerializedName("data")
-	val productDetailsDM: ProductDetailsDM? = null
+    @field:SerializedName("data")
+    val productDetailsDM: ProductDetailsDM? = null
 )
 
 data class ProductDetailsBrandDM(
 
-	@field:SerializedName("image")
-	val image: String? = null,
+    @field:SerializedName("image")
+    val image: String? = null,
 
-	@field:SerializedName("name")
-	val name: String? = null,
+    @field:SerializedName("name")
+    val name: String? = null,
 
-	@field:SerializedName("_id")
-	val id: String? = null,
+    @field:SerializedName("_id")
+    val id: String? = null,
 
-	@field:SerializedName("slug")
-	val slug: String? = null
+    @field:SerializedName("slug")
+    val slug: String? = null
 )
 
 data class ProductDetailsDM(
 
-	@field:SerializedName("sold")
-	val sold: Int? = null,
+    @field:SerializedName("sold")
+    val sold: Int? = null,
 
-	@field:SerializedName("images")
-	val images: List<String>? = null,
+    @field:SerializedName("images")
+    val images: List<String>? = null,
 
-	@field:SerializedName("quantity")
-	val quantity: Int? = null,
+    @field:SerializedName("quantity")
+    val quantity: Int? = null,
 
-	@field:SerializedName("imageCover")
-	val imageCover: String? = null,
+    @field:SerializedName("imageCover")
+    val imageCover: String? = null,
 
-	@field:SerializedName("description")
-	val description: String? = null,
+    @field:SerializedName("description")
+    val description: String? = null,
 
-	@field:SerializedName("title")
-	val title: String? = null,
+    @field:SerializedName("title")
+    val title: String? = null,
 
-	@field:SerializedName("ratingsQuantity")
-	val ratingsQuantity: Int? = null,
+    @field:SerializedName("ratingsQuantity")
+    val ratingsQuantity: Int? = null,
 
-	@field:SerializedName("ratingsAverage")
-	val ratingsAverage: Int? = null,
+    @field:SerializedName("ratingsAverage")
+    val ratingsAverage: Double? = null,
 
-	@field:SerializedName("createdAt")
-	val createdAt: String? = null,
+    @field:SerializedName("createdAt")
+    val createdAt: String? = null,
 
-	@field:SerializedName("price")
-	val price: Int? = null,
+    @field:SerializedName("price")
+    val price: Int? = null,
 
-	@field:SerializedName("__v")
-	val v: Int? = null,
+    @field:SerializedName("__v")
+    val v: Int? = null,
 
-	@field:SerializedName("_id")
-	val id: String? = null,
+    @field:SerializedName("_id")
+    val id: String? = null,
 
-	@field:SerializedName("brand")
-	val productDetailsBrandDM: ProductDetailsBrandDM? = null,
+    @field:SerializedName("brand")
+    val productDetailsBrandDM: ProductDetailsBrandDM? = null,
 
-	@field:SerializedName("slug")
-	val slug: String? = null,
+    @field:SerializedName("slug")
+    val slug: String? = null,
 
-	@field:SerializedName("updatedAt")
-	val updatedAt: String? = null
+    @field:SerializedName("updatedAt")
+    val updatedAt: String? = null
 )
